+++ dwds/lib/src/debugging/debugger.dart
@@ -206,8 +206,9 @@ class Debugger extends Domain {
         ..tokenPos = location.tokenPos);
     _streamNotify(
         'Debug',
-        Event()
-          ..kind = EventKind.kBreakpointAdded
+        Event(
+            kind: EventKind.kBreakpointAdded,
+            timestamp: DateTime.now().millisecondsSinceEpoch)
           ..isolate = inspector.isolateRef
           ..breakpoint = breakpoint);
     return breakpoint;
@@ -228,8 +229,9 @@ class Debugger extends Domain {
     }
     _streamNotify(
         'Debug',
-        Event()
-          ..kind = EventKind.kBreakpointRemoved
+        Event(
+            kind: EventKind.kBreakpointRemoved,
+            timestamp: DateTime.now().millisecondsSinceEpoch)
           ..isolate = inspector.isolateRef
           ..breakpoint = bp);
     await _removeBreakpoint(jsId);
@@ -389,21 +391,23 @@ class Debugger extends Domain {
   Future<void> _pauseHandler(DebuggerPausedEvent e) async {
     var isolate = inspector.isolate;
     if (isolate == null) return;
-    var event = Event()..isolate = inspector.isolateRef;
+    Event event;
+    var timestamp = DateTime.now().millisecondsSinceEpoch;
     var params = e.params;
     var breakpoints = params['hitBreakpoints'] as List;
     if (breakpoints.isNotEmpty) {
-      event.kind = EventKind.kPauseBreakpoint;
+      event = Event(kind: EventKind.kPauseBreakpoint, timestamp: timestamp);
     } else if (e.reason == 'exception' || e.reason == 'assert') {
-      event.kind = EventKind.kPauseException;
+      event = Event(kind: EventKind.kPauseException, timestamp: timestamp);
     } else {
       // If we don't have source location continue stepping.
       if (_isStepping && _sourceLocation(e) == null) {
         await _remoteDebugger.sendCommand('Debugger.stepInto');
         return;
       }
-      event.kind = EventKind.kPauseInterrupted;
+      event = Event(kind: EventKind.kPauseInterrupted, timestamp: timestamp);
     }
+    event.isolate = inspector.isolateRef;
     var jsFrames =
         (e.params['callFrames'] as List).cast<Map<String, dynamic>>();
     var frames = await dartFramesFor(jsFrames);
@@ -424,8 +428,9 @@ class Debugger extends Domain {
     if (isolate == null) return;
     _pausedStack = null;
     _pausedJsStack = null;
-    var event = Event()
-      ..kind = EventKind.kResume
+    var event = Event(
+        kind: EventKind.kResume,
+        timestamp: DateTime.now().millisecondsSinceEpoch)
       ..isolate = inspector.isolateRef;
     isolate.pauseEvent = event;
     _streamNotify('Debug', event);